aider: The logging level has been updated to INFO and a new formatter has been set for the log messages to include a timestamp, the log level, and the message, while removing duplicate logging output.
# Aider chat conversation:

USER: can you help make logs more clear and concise? '2023-12-20 08:00:04.051 [ttb] [2023-12-20 13:00:04,051] INFO in app: Processed update message for symbol: USTEC100
2023-12-20 08:00:04.051 [ttb] Processed update message for symbol: USTEC100
2023-12-20 08:00:04.051 [ttb] 10.207.127.242 - - [20/Dec/2023 13:00:04] "POST /webhook HTTP/1.1" 200 -
2023-12-20 08:00:04.367 [ttb] [2023-12-20 13:00:04,367] DEBUG in app: Received webhook data: type=update,symbol=AUDJPY,keyword=TD9buy
2023-12-20 08:00:04.367 [ttb] Received webhook data: type=update,symbol=AUDJPY,keyword=TD9buy
2023-12-20 08:00:04.680 [ttb] [2023-12-20 13:00:04,679] DEBUG in app: Received webhook data: type=update,symbol=EURJPY,keyword=TD9buyOFF
2023-12-20 08:00:04.680 [ttb] Received webhook data: type=update,symbol=EURJPY,keyword=TD9buyOFF
2023-12-20 08:00:04.798 [ttb] [2023-12-20 13:00:04,798] INFO in app: Processed update message for symbol: AUDJPY
2023-12-20 08:00:04.798 [ttb] Processed update message for symbol: AUDJPY
2023-12-20 08:00:04.798 [ttb] 10.207.169.21 - - [20/Dec/2023 13:00:04] "POST /webhook HTTP/1.1" 200 -
2023-12-20 08:00:05.155 [ttb] [2023-12-20 13:00:05,155] INFO in app: Processed update message for symbol: EURJPY
2023-12-20 08:00:05.155 [ttb] Processed update message for symbol: EURJPY
2023-12-20 08:00:05.156 [ttb] 10.207.169.23 - - [20/Dec/2023 13:00:05] "POST /webhook HTTP/1.1" 200 -

ASSISTANT: To make the logs more clear and concise, we can adjust the logging format to remove redundant information and ensure that each log entry is on a single line. We will modify the logging configuration in `app.py` to achieve this.

Here are the diffs for those changes:

```diff
--- app.py
+++ app.py
@@ ... @@
 app = Flask(__name__)
 handler = logging.StreamHandler()
-handler.setLevel(logging.DEBUG)
+handler.setLevel(logging.INFO)  # Set to INFO to reduce the verbosity of the logs
 formatter = logging.Formatter('%(asctime)s [%(levelname)s] %(message)s', datefmt='%Y-%m-%d %H:%M:%S')
 handler.setFormatter(formatter)
 app.logger.addHandler(handler)
-app.logger.setLevel(logging.DEBUG)
+app.logger.setLevel(logging.INFO)  # Set to INFO to reduce the verbosity of the logs
 app.logger.propagate = False

 class AirtableOperations:
@@ ... @@
```

This change updates the logging level to
^C KeyboardInterrupt

USER: can you try again?

ASSISTANT: To make the logs more clear and concise, we will adjust the logging configuration in `app.py` to use a simpler format and to log at the INFO level by default, which will reduce the verbosity of the logs. We will also remove the duplicate timestamp and logging level that is currently being printed twice.

Here are the diffs for those changes:

```diff
--- app.py
+++ app.py
@@ -6,11 +6,13 @@
 app = Flask(__name__)
 handler = logging.StreamHandler()
-handler.setLevel(logging.DEBUG)
+handler.setLevel(logging.INFO)  # Set to INFO to reduce the verbosity of the logs
-app.logger.addHandler(handler)
-app.logger.setLevel(logging.DEBUG)
+formatter = logging.Formatter('%(asctime)s [%(levelname)s] %(message)s', datefmt='%Y-%m-%d %H:%M:%S')
+handler.setFormatter(formatter)
+app.logger.addHandler(handler)
+app.logger.setLevel(logging.INFO)  # Set to INFO to reduce the verbosity of the logs
 app.logger.propagate = False

 class AirtableOperations:
```

This change updates the logging level to INFO and sets a new formatter for the log messages to include a timestamp, the log level, and the message. It also specifies a date format that does not include the milliseconds, which makes the logs cleaner. The duplicate logging output is removed by not printing the log level and timestamp twice.